+++ src/eve-server/character/CharacterDB.h
@@ -187,7 +187,7 @@ public:
      */
     bool        SaveSkillQueue(uint32 charID, SkillQueue &queue);
     bool        SavePausedSkillQueue(uint32 charID, SkillQueue &queue);
-    void        SaveSkillHistory(uint16 eventID, double logDate, uint32 characterID, uint32 skillTypeID, uint8 skillLevel, uint32 absolutePoints);
+    void        SaveSkillHistory( uint16 eventID, int64 logDate, uint32 characterID, uint32 skillTypeID, uint8 skillLevel, uint32 absolutePoints );
     PyRep*      GetSkillHistory(uint32 charID);
     void        UpdateSkillQueueEndTime(int64 endtime, uint32 charID);
